Act 4 Fixes
Found out that Heart fight music was not always playing and while troubleshooting, found the Shield and Spear fight wasn't turning back on the music when going back above health.

Changed Heart fight music from a background music call to a tempbackgroundmusic call which is more consistent.
Added a new check for Shield and Spear fight to make sure it properly keeps the music instead of reverting to normal music.